requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-groq==0.1.2
pypdf==3.17.4
chromadb==0.3.13
sentence-transformers==2.2.2
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
tiktoken==0.5.2
numpy==1.24.3
pandas==2.0.3
pyarrow==12.0.1
transformers==4.30.2
huggingface_hub==0.16.4
altair==4.2.2
pydantic==1.10.13 